itemDB 수정 + controller, dao, service, vo 수정 mapper , action-mybatis 수정
