+++ vampir/src/main/java/com/myspring/vampir/item/controller/ItemController.java
@@ -7,5 +7,5 @@ import org.springframework.web.servlet.ModelAndView;
 
 public interface ItemController {
 	public ModelAndView listItems(HttpServletRequest request, HttpServletResponse response) throws Exception;
-	
+
 }
