+++ vampir/src/main/java/com/myspring/vampir/item/dao/ItemDAO.java
@@ -1,9 +1,10 @@
 package com.myspring.vampir.item.dao;
 
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.dao.DataAccessException;
 
 public interface ItemDAO {
-	public List selectAllItemList() throws DataAccessException;
+	List<Map<String,Object>> selectAllUnified() throws DataAccessException;
 }
