+++ vampir/src/main/java/com/myspring/vampir/item/service/ItemServiceImpl.java
@@ -1,26 +1,21 @@
 package com.myspring.vampir.item.service;
 
 import java.util.List;
+import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
 
 import com.myspring.vampir.item.dao.ItemDAO;
 
 @Service("itemService")
-@Transactional(propagation = Propagation.REQUIRED)
 public class ItemServiceImpl implements ItemService {
+	
 	@Autowired
 	private ItemDAO itemDAO;
 	
 	@Override
-	public List listItems() throws DataAccessException {
-		List itemsList = null;
-		itemsList = itemDAO.selectAllItemList();
-		return itemsList;
+	public List<Map<String, Object>> listItemsUnified() {
+		return itemDAO.selectAllUnified();
 	}
-	
 }
