+++ vampir/src/main/java/com/myspring/vampir/item/vo/AccessoryVO.java
@@ -1,0 +1,16 @@
+package com.myspring.vampir.item.vo;
+
+import org.springframework.stereotype.Component;
+
+@Component("AccessoryVO")
+public class AccessoryVO {
+	private String item_code;
+	
+	public AccessoryVO() {
+		
+	}
+	
+	public AccessoryVO(String item_code) {
+		this.item_code = item_code;
+	}
+}
