+++ vampir/src/main/java/com/myspring/vampir/item/vo/SubequipVO.java
@@ -1,0 +1,16 @@
+package com.myspring.vampir.item.vo;
+
+import org.springframework.stereotype.Component;
+
+@Component("SubequipVO")
+public class SubequipVO {
+	private String item_code;
+	
+	public SubequipVO() {
+		
+	}
+	
+	public SubequipVO(String item_code) {
+		this.item_code = item_code;
+	}
+}
